Solution to Input a Character and Print Its ASCII Value

+++ Input a Character and Print Its ASCII Value@@6788d5012db25efec946f93f/code.c
@@ -5,6 +5,6 @@ int main()
     char ch;
     printf("");
     scanf("%c", &ch);
-    printf("ASCII Value: %d", ch, (int)ch);
+    printf("ASCII Value: %d", ch, ch);
     return 0;
 }
